分布式事务学习--seata

+++ distributed-transaction/dtx-seata-demo/dtx-seata-demo-bank1/src/main/java/com/deng/bank1/tx/service/BankService.java
@@ -2,5 +2,5 @@ package com.deng.bank1.tx.service;
 
 
 public interface BankService {
-    boolean transfer(Long id,Double amount);
+    boolean transfer(Integer id,Long amount);
 }
